requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.3
Babel==2.8.0
beautifulsoup4==4.9.1
bleach==3.1.5
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
distlib==0.3.1
distro==1.5.0
Django==3.0.8
django-appconf==1.0.4
django-autofixture==0.12.1
django-bootstrap-form==3.4
django-bootstrap4==2.2.0
django-braces==1.14.0
django-cachalot==2.2.2
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cleanup==5.0.0
django-comments-dab==1.6.7
django-compat==1.0.15
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-friendship @ git+https://github.com/HiveTechies/django-friendship
django-hitcount @ git+https://github.com/HiveTechies/django-hitcount
django-imagekit==4.0.2
django-ipware==3.0.0
django-js-asset==1.2.2
django-language-field==0.0.3
django-maintenance-mode @ git+https://github.com/HiveTechies/django-maintenance-mode-changed.git@ad6016c15347957b5565255419f5bf11d9b51604
django-markdown2==0.3.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-nyt==1.1.5
django-postman @ git+https://github.com/HiveTechies/postman-changed.git@2e50ab3c76c35c65f476055417334a92252b5933
django-qrcode==0.3
django-ranged-response==0.2.0
django-reversion==3.0.7
django-sekizai==1.1.0
django-star-ratings==0.9.1
django-taggit==1.3.0
django-tellme==0.6.5
django-templated-mail==1.1.1
django-termsandconditions==2.0.6
django-tinymce==3.0.2
django-tracking @ git+https://github.com/hivetechies/django-tracking
django-tracking-analyzer==1.1.1
django-user-accounts==3.0.2
django-user-agents==0.4.0
django-xmlrpc==0.1.8
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.16
easy-thumbnails==2.7
environ==1.0
geoip2==3.0.0
gunicorn==20.0.4
html5lib==1.1
icalendar==4.0.6
idna==2.10
imagesize==1.2.0
ipaddr==2.2.0
Jinja2==2.11.2
lockfile==0.12.2
Markdown==2.6.11
markdown2==2.3.9
MarkupSafe==1.1.1
maxminddb==1.5.4
mots-vides==2015.5.11
msgpack==1.0.0
packaging==20.4
pep517==0.8.2
pilkit==2.0
Pillow==7.2.0
pinax-calendars==2.0.4
pinax-images==4.0.0
pinax-invitations==6.1.2
pinax-teams @ git+https://github.com/HiveTechies/pinax-teams-changed.git@2a4e47d7febb709c122b9589ac3cdea2af4ee37e
pinax-templates==2.0.6
progress==1.5
psycopg2==2.8.5
progress==1.5
pycodestyle==2.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytoml==0.1.21
pytz==2020.1
regex==2020.6.8
requests==2.24.0
retrying==1.3.3
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
Sphinx==3.1.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.1
ua-parser==0.10.0
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.9
user-agents==2.1
webencodings==0.5.1
wiki==0.6
whitenoise
django-heroku
git+https://github.com/HiveTechies/todo-changed
django-admin-honeypot
